IOS-8933 Moved the brand font weights palettes to different files. Changed the MisticaFontWeights protocol to be consistent with the naming and have a pattern to be reproduced from the templates.
