+++ Sources/MisticaCommon/Fonts/Brands/BlauFontWeights.swift
@@ -1,0 +1,19 @@
+//
+//  BlauFontWeights.swift
+//
+//  Made with ❤️ by Novum
+//
+//  Copyright © Telefonica. All rights reserved.
+//
+
+import Foundation
+
+struct BlauFontWeights: MisticaFontWeights {
+    public var textCardtitle: MisticaFontWeightType = .regular
+    public var textPreset5: MisticaFontWeightType = .light
+    public var textPreset6: MisticaFontWeightType = .light
+    public var textPreset7: MisticaFontWeightType = .light
+    public var textPreset8: MisticaFontWeightType = .light
+    public var textPreset9: MisticaFontWeightType = .light
+    public var textPreset10: MisticaFontWeightType = .light
+}
