+++ Sources/MisticaCommon/Fonts/Brands/MovistarFontWeights.swift
@@ -1,0 +1,19 @@
+//
+//  MovistarFontWeights.swift
+//
+//  Made with ❤️ by Novum
+//
+//  Copyright © Telefonica. All rights reserved.
+//
+
+import Foundation
+
+struct MovistarFontWeights: MisticaFontWeights {
+    public var textCardtitle: MisticaFontWeightType = .bold
+    public var textPreset5: MisticaFontWeightType = .bold
+    public var textPreset6: MisticaFontWeightType = .bold
+    public var textPreset7: MisticaFontWeightType = .bold
+    public var textPreset8: MisticaFontWeightType = .bold
+    public var textPreset9: MisticaFontWeightType = .bold
+    public var textPreset10: MisticaFontWeightType = .bold
+}
