+++ Sources/MisticaCommon/Fonts/FontStyle+Weights.swift
@@ -52,48 +52,8 @@ public extension FontStyle {
             case .light: return .light
             case .regular: return .regular
             case .medium: return .medium
-            case .cardTitle: return MisticaConfig.currentFontWeights.cardTitlePreset
+            case .cardTitle: return MisticaConfig.currentFontWeights.textCardtitle
             }
         }
     }
-}
-
-struct MovistarFontWeights: MisticaFontWeights {
-    public var cardTitlePreset: MisticaFontWeightType = .bold
-    public var textPreset5: MisticaFontWeightType = .bold
-    public var textPreset6: MisticaFontWeightType = .bold
-    public var textPreset7: MisticaFontWeightType = .bold
-    public var textPreset8: MisticaFontWeightType = .bold
-    public var textPreset9: MisticaFontWeightType = .bold
-    public var textPreset10: MisticaFontWeightType = .bold
-}
-
-struct O2FontWeights: MisticaFontWeights {
-    public var cardTitlePreset: MisticaFontWeightType = .regular
-    public var textPreset5: MisticaFontWeightType = .light
-    public var textPreset6: MisticaFontWeightType = .light
-    public var textPreset7: MisticaFontWeightType = .light
-    public var textPreset8: MisticaFontWeightType = .light
-    public var textPreset9: MisticaFontWeightType = .light
-    public var textPreset10: MisticaFontWeightType = .light
-}
-
-struct BlauFontWeights: MisticaFontWeights {
-    public var cardTitlePreset: MisticaFontWeightType = .regular
-    public var textPreset5: MisticaFontWeightType = .light
-    public var textPreset6: MisticaFontWeightType = .light
-    public var textPreset7: MisticaFontWeightType = .light
-    public var textPreset8: MisticaFontWeightType = .light
-    public var textPreset9: MisticaFontWeightType = .light
-    public var textPreset10: MisticaFontWeightType = .light
-}
-
-struct VivoFontWeights: MisticaFontWeights {
-    public var cardTitlePreset: MisticaFontWeightType = .regular
-    public var textPreset5: MisticaFontWeightType = .light
-    public var textPreset6: MisticaFontWeightType = .light
-    public var textPreset7: MisticaFontWeightType = .light
-    public var textPreset8: MisticaFontWeightType = .light
-    public var textPreset9: MisticaFontWeightType = .light
-    public var textPreset10: MisticaFontWeightType = .light
 }
